pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hybrid-db"
version = "0.1.0"
description = "Local vector + graph native database prototype"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.95.2",
    "uvicorn==0.24.0.post1",
    "typer==0.12.3",
    "numpy==2.3.5",
    "pydantic==1.10.15",
    "rich==13.7.1",
    "faiss-cpu==1.13.0",
    "sentence-transformers==2.7.0",
    "networkx==3.2.1",
]

[tool.setuptools.packages.find]
where = ["src"]

